+++ src/settings/tokens/languages/diff.ts
@@ -1,4 +1,4 @@
-import { Colors } from '../../../themes/Theme';
+import { Colors } from '../../Theme';
 import TokenSettings, { TokenFontStyle } from '../TokenSettings';
 
 export default ({ base, misc }: Colors): TokenSettings[] => {
