+++ src/settings/tokens/languages/php.ts
@@ -1,4 +1,4 @@
-import { Colors } from '../../../themes/Theme';
+import { Colors } from '../../Theme';
 import TokenSettings from '../TokenSettings';
 
 export default ({ base }: Colors): TokenSettings[] => [
